+++ packages/components/button/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sl-design-system/button
 
+## 0.0.15
+
+### Patch Changes
+
+- [#795](https://github.com/sl-design-system/components/pull/795) [`c15e60e`](https://github.com/sl-design-system/components/commit/c15e60eef61430b3a9b944c4f8438d19ce8bb3c1) - Fix background for default buttons + fix hover state for all buttons
+
+- [#724](https://github.com/sl-design-system/components/pull/724) [`b3f9d42`](https://github.com/sl-design-system/components/commit/b3f9d42945c7b427105353ede3cf74ba3191792d) - Added avatar + added token for focus ring offset
+
+- Updated dependencies [[`198c7c2`](https://github.com/sl-design-system/components/commit/198c7c277787d48276d402522f94f2d0d5132152), [`b3f9d42`](https://github.com/sl-design-system/components/commit/b3f9d42945c7b427105353ede3cf74ba3191792d)]:
+  - @sl-design-system/shared@0.2.0
+
 ## 0.0.14
 
 ### Patch Changes