+++ packages/components/icon/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sl-design-system/icon
 
+## 0.0.5
+
+### Patch Changes
+
+- [#786](https://github.com/sl-design-system/components/pull/786) [`4ca7b20`](https://github.com/sl-design-system/components/commit/4ca7b20ee7d09ee2ccfcf2743fd48f00a8207e39) - Fix outline being shown when focused
+
+- [#786](https://github.com/sl-design-system/components/pull/786) [`4ca7b20`](https://github.com/sl-design-system/components/commit/4ca7b20ee7d09ee2ccfcf2743fd48f00a8207e39) - Remove tabindex=-1 which causes the icon to receive focus
+
+- [#790](https://github.com/sl-design-system/components/pull/790) [`9b8e371`](https://github.com/sl-design-system/components/commit/9b8e371932fbe979f3250e07c605ad39239d4f82) - Update to the latest fontawesome icons
+
 ## 0.0.4
 
 ### Patch Changes