+++ packages/themes/nowa-era/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sl-design-system/nowa-era
 
+## 0.0.4
+
+### Patch Changes
+
+- [#755](https://github.com/sl-design-system/components/pull/755) [`b657079`](https://github.com/sl-design-system/components/commit/b657079044ed53a9b49d4374a08bdc3a32179eeb) - Added tokens for popover
+
+- [#724](https://github.com/sl-design-system/components/pull/724) [`b3f9d42`](https://github.com/sl-design-system/components/commit/b3f9d42945c7b427105353ede3cf74ba3191792d) - Added avatar + added token for focus ring offset
+
+- Updated dependencies [[`4ca7b20`](https://github.com/sl-design-system/components/commit/4ca7b20ee7d09ee2ccfcf2743fd48f00a8207e39), [`4ca7b20`](https://github.com/sl-design-system/components/commit/4ca7b20ee7d09ee2ccfcf2743fd48f00a8207e39), [`9b8e371`](https://github.com/sl-design-system/components/commit/9b8e371932fbe979f3250e07c605ad39239d4f82)]:
+  - @sl-design-system/icon@0.0.5
+
 ## 0.0.3
 
 ### Patch Changes